requirements.txt
torch>=2.0.0
torchvision>=0.15.0
torchgeo>=0.5.0
lightning>=2.0.0
tensorboard>=2.13.0
matplotlib>=3.7.0
numpy>=1.24.0
pandas>=2.0.0
pydantic>=2.0.0
pyyaml>=6.0
pillow>=9.5.0
scikit-learn>=1.2.0
seaborn>=0.12.0
tqdm>=4.65.0
optuna>=3.1.0
kornia>=0.7.0
python-dotenv>=1.0.0